No Primary Up and Make the Secondary as Primary
Continue steps After Installation_Step_Of_1_Primary_2_Secondary_Mongodb_On_Centos_8.sql
If you have no primary node up and need to make the secondary node as the primary node in MongoDB. 
This is a tricky situation because normally you need a majority of the replica set members to be available 
to elect a new primary node. However, there are some possible ways to force a secondary node to become the primary node in this case.